add new file export_utils4.c and modified some files
