+++ built_in/error.c
@@ -6,7 +6,7 @@
 /*   By: junseyun <junseyun@student.42gyeongsan.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/06 15:09:01 by junseyun          #+#    #+#             */
-/*   Updated: 2024/11/24 17:29:33 by junseyun         ###   ########.fr       */
+/*   Updated: 2024/11/24 20:05:23 by junseyun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,6 +27,6 @@ int	print_error(int i)
 
 int	print_export_error(char *str)
 {
-	printf("mini: export : \'%s\': not a valid identifier\n");
+	printf("mini: export : \'%s\': not a valid identifier\n", str);
 	return (1);
 }
